feat: run migrations on startup
`fisherman` supported running pre-commands before the binary started up,
allowing us to use `sqlx migrate run` beforehand. Kubernetes doesn't
support this, so we need to run the migrations when the service starts
up instead.
